fix-aboutページをコンポーネントに切り出した
